+++ service/resource/configmapv1/delete_test.go
@@ -1,4 +1,4 @@
-package configmap
+package configmapv1
 
 import (
 	"context"
