+++ service/resource/configmapv1/desired.go
@@ -1,12 +1,11 @@
-package service
+package configmapv1
 
 import (
 	"context"
 	"fmt"
+	"strconv"
 
 	"github.com/giantswarm/microerror"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	apiv1 "k8s.io/client-go/pkg/api/v1"
 )
 
 func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interface{}, error) {
@@ -17,26 +16,19 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interf
 
 	r.logger.Log("cluster", customObject.Spec.GuestCluster.ID, "debug", "get desired state")
 
-	// Lookup the desired state of the service to have a reference of ports how
-	// they should be.
-	dState := []apiv1.ServicePort{}
+	// Lookup the desired state of the config map to have a reference of data how
+	// it should be.
+	dState := map[string]string{}
 	for _, p := range customObject.Spec.ProtocolPorts {
-		servicePortName := fmt.Sprintf(
-			PortNameFormat,
-			p.Protocol,
+		configMapKey := strconv.Itoa(p.LBPort)
+		configMapValue := fmt.Sprintf(
+			DataValueFormat,
+			customObject.Spec.GuestCluster.Namespace,
+			customObject.Spec.GuestCluster.Service,
 			p.IngressPort,
-			customObject.Spec.GuestCluster.ID,
 		)
 
-		newPort := apiv1.ServicePort{
-			Name:       servicePortName,
-			Protocol:   apiv1.ProtocolTCP,
-			Port:       int32(p.LBPort),
-			TargetPort: intstr.FromInt(p.LBPort),
-			NodePort:   int32(p.LBPort),
-		}
-
-		dState = append(dState, newPort)
+		dState[configMapKey] = configMapValue
 	}
 
 	r.logger.Log("cluster", customObject.Spec.GuestCluster.ID, "debug", fmt.Sprintf("found desired state: %#v", dState))
