+++ service/resource/configmapv1/error.go
@@ -1,4 +1,4 @@
-package service
+package configmapv1
 
 import (
 	"github.com/giantswarm/microerror"
@@ -9,13 +9,6 @@ var invalidConfigError = microerror.New("invalid config")
 // IsInvalidConfig asserts invalidConfigError.
 func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
-}
-
-var servicePortNotFoundError = microerror.New("service port not found")
-
-// IsServicePortNotFound asserts servicePortNotFoundError.
-func IsServicePortNotFound(err error) bool {
-	return microerror.Cause(err) == servicePortNotFoundError
 }
 
 var wrongTypeError = microerror.New("wrong type")
