+++ service/resource/configmapv1/resource.go
@@ -1,4 +1,4 @@
-package configmap
+package configmapv1
 
 import (
 	"github.com/giantswarm/ingresstpr"
