+++ service/resource/servicev1/resource.go
@@ -1,4 +1,4 @@
-package service
+package servicev1
 
 import (
 	"github.com/giantswarm/ingresstpr"
